clients/gpslogger.c: Address Coverity CID 355859
Coverity is worried about truncating a time_t  to an unsigned int to
pass to sleep().  How long do they think gposlogger could be asked to
wait?

+++ clients/gpxlogger.c
@@ -469,7 +469,8 @@ int main(int argc, char **argv)
             break;
         }
         // avoid banging on reconnect
-        (void)sleep(timeout);
+        // (unsigned int) cast to shut up Coverity CID 355859
+        (void)sleep((unsigned int)timeout);
         syslog(LOG_INFO, "timeout; about to reconnect");
     }
